+++ src/DependencyInjection/Compiler/AlertProviderPass.php
@@ -1,0 +1,27 @@
+<?php
+/*
+ * (c) 2026: 975L <contact@975l.com>
+ * (c) 2026: Laurent Marquet <laurent.marquet@laposte.net>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace c975L\ConfigBundle\DependencyInjection\Compiler;
+
+use c975L\ConfigBundle\Management\AlertProviderInterface;
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+
+class AlertProviderPass implements CompilerPassInterface
+{
+    public function process(ContainerBuilder $container): void
+    {
+        foreach ($container->getDefinitions() as $id => $definition) {
+            $class = $definition->getClass();
+            if ($class && is_subclass_of($class, AlertProviderInterface::class)) {
+                $definition->addTag('c975l.alert_provider');
+            }
+        }
+    }
+}